增加redis重连的机制的显式设置以及给予重连事件的监听事件
