+++ FastCache.Core/Driver/IRedisCache.cs
@@ -10,6 +10,8 @@ namespace FastCache.Core.Driver
 {
     public interface IRedisCache : ICacheClient
     {
+        void Dispose();
+        
         ConnectionMultiplexer GetConnectionMultiplexer();
 
         RedLockFactory GetRedLockFactory();
